requirements.txt
numpy>=1.18.4
pandas==1.3.5
pyodbc>=4.0.30
python-dotenv>=0.13.0
python3-logstash>=0.4.80
pytz>=2020.1
alive-progress==1.5.1
configargparse>=1.2.3
loguru>=0.5.3
flask>=1.1.2
psutil>=5.8.0
Flask-HTTPAuth>=4.4.0
psycopg2>=2.9.1
licenseheaders>=0.8.8
APScheduler>=3.7.0
openpyxl>=3.0.9
requests>=2.26.0
scikit-learn>=1.0.1
matplotlib>=3.5.1
